Fixing rounding problems in piping. WTI-2875

+++ Riskeer/Piping/src/Riskeer.Piping.Data/DerivedPipingInput.cs
@@ -388,7 +388,7 @@ namespace Riskeer.Piping.Data
                 double thicknessUnderSurface = Math.Min(layer.Top, surfaceLevel) - bottom;
 
                 totalThickness += thicknessUnderSurface;
-                weighedTotal += layer.BelowPhreaticLevel.Mean * thicknessUnderSurface;
+                weighedTotal += (double) layer.BelowPhreaticLevel.Mean * thicknessUnderSurface;
             }
 
             return weighedTotal / totalThickness;
@@ -408,7 +408,7 @@ namespace Riskeer.Piping.Data
                 double thicknessUnderSurface = Math.Min(layer.Top, surfaceLevel) - bottom;
 
                 totalThickness += thicknessUnderSurface;
-                weighedTotal += layer.Permeability.Mean * thicknessUnderSurface;
+                weighedTotal += (double) layer.Permeability.Mean * thicknessUnderSurface;
             }
 
             return weighedTotal / totalThickness;
